backend/pyproject.toml
[project]
name = "ecommerce-backend"
version = "1.0.0"
description = "Professional E-commerce Backend API"
requires-python = ">=3.11"
dependencies = [
    "django>=4.2.0",
    "djangorestframework>=3.14.0",
    "psycopg2-binary>=2.9.0",
    "redis>=4.5.0",
    "celery>=5.3.0",
    "pillow>=10.0.0",
    "stripe>=5.5.0",
    "python-decouple>=3.8",
    "django-cors-headers>=4.2.0",
    "drf-spectacular>=0.26.0",
    "dj-database-url>=3.0.1",
    "django-ratelimit>=4.1.0",
    "bleach>=6.1.0",
    "django-redis>=5.4.0",
    "gunicorn>=21.2.0",
    "djangorestframework-simplejwt>=5.5.1",
    "django-otp>=1.6.1",
    "qrcode>=8.2",
    "django-phonenumber-field>=8.1.0",
    "phonenumbers>=9.0.15",
]

[tool.uv]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.5.0", 
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.3.0",
    "black>=23.7.0",
    "commitizen>=3.6.0",
    "django-debug-toolbar>=4.2.0",
    "factory-boy>=3.3.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "PIE", "T20"]
ignore = ["E501", "N806", "N803"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"**/migrations/*" = ["E501", "F401"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
exclude = ["migrations/"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.testing"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["apps/"]
addopts = "--cov=apps --cov-report=html --cov-report=term-missing --reuse-db"

[tool.commitizen]
name = "cz_conventional_commits"
version = "1.0.0"
tag_format = "v$version"
